# Conflicts:
#	back/CC.Master.sln

+++ back/Instances/Instances.Application/Demos/Deletion/InactiveDemosCleaner.cs
@@ -1,9 +1,12 @@
-﻿using Email.Domain;
+using Email.Domain;
 using Instances.Application.Demos.Emails;
 using Instances.Domain.Demos;
 using Instances.Domain.Demos.Cleanup;
+using Instances.Domain.Demos.Filtering;
 using Instances.Domain.Instances;
 using Instances.Domain.Shared;
+using MoreLinq.Extensions;
+using Rights.Domain.Filtering;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -12,8 +15,16 @@ using Tools;
 
 namespace Instances.Application.Demos.Deletion
 {
+    public class DemoCleanupParams
+    {
+        public bool IsDryRun { get; set; }
+    }
+
     public class InactiveDemosCleaner
     {
+        private const int MaxConcurrentDemoUsageRetrievalAttempts = 200;
+        private static readonly TimeSpan DelayBetweenConcurrentDemoUsageRetrievalAttempts = TimeSpan.FromMilliseconds(500);
+
         private readonly ITimeProvider _timeProvider;
         private readonly IInstanceSessionLogsService _sessionLogsService;
         private readonly IDemosStore _demosStore;
@@ -21,6 +32,7 @@ namespace Instances.Application.Demos.Deletion
         private readonly IEmailService _emailService;
         private readonly IDemoEmails _demoEmails;
         private readonly IDemoDeletionCalculator _deletionCalculator;
+        private readonly IInstancesStore _instancesStore;
 
         public InactiveDemosCleaner
         (
@@ -30,7 +42,8 @@ namespace Instances.Application.Demos.Deletion
             ICcDataService ccDataService,
             IEmailService emailService,
             IDemoEmails demoEmails,
-            IDemoDeletionCalculator deletionCalculator
+            IDemoDeletionCalculator deletionCalculator,
+            IInstancesStore instancesStore
         )
         {
             _timeProvider = timeProvider;
@@ -40,20 +53,65 @@ namespace Instances.Application.Demos.Deletion
             _emailService = emailService;
             _demoEmails = demoEmails;
             _deletionCalculator = deletionCalculator;
+            _instancesStore = instancesStore;
         }
 
-        public async Task CleanAsync()
+        public async Task CleanAsync(DemoCleanupParams demoCleanupParams)
         {
+            var eligibleDemos = new DemoFilter
+            {
+                IsActive = BoolCombination.TrueOnly,
+                IsTemplate = BoolCombination.FalseOnly,
+                IsProtected = BoolCombination.FalseOnly
+            };
+            var usages = await GetDemoUsagesAsync(eligibleDemos);
+            await UpdateDemosAsync(usages);
+
             var today = _timeProvider.Today();
-            var activeDemos = await _demosStore.GetActiveAsync();
-            var infoTasks = activeDemos
-                .Where(d => !d.IsTemplate && !d.Instance.IsProtected)
-                .Select(d => GetUpdatedCleanupInfoAsync(d, today));
+            var infos = await GetCleanupInfoAsync(usages, today);
+            await ReportCleanupIntentionsAsync(infos, demoCleanupParams.IsDryRun);
 
-            var infos = await Task.WhenAll(infoTasks);
+            if (demoCleanupParams.IsDryRun)
+            {
+                return;
+            }
 
-            await ReportCleanupIntentionsAsync(infos.Where(i => i.ShouldBeReported()));
             await DeleteAllAsync(infos.Where(i => i.State == DemoState.DeletionScheduledToday));
+        }
+
+        private async Task<List<DemoLastUsageRetrieveAttempt>> GetDemoUsagesAsync(DemoFilter filter)
+        {
+            var result = new List<DemoLastUsageRetrieveAttempt>();
+            var activeDemos = await _demosStore.GetAsync(filter, AccessRight.All);
+
+            foreach (var batch in activeDemos.Batch(MaxConcurrentDemoUsageRetrievalAttempts))
+            {
+                result.AddRange(await GetDemoUsagesWithoutHarassingPlatformAsync(batch));
+            }
+            return result;
+        }
+
+        private async Task<DemoLastUsageRetrieveAttempt[]> GetDemoUsagesWithoutHarassingPlatformAsync(IEnumerable<Demo> demos)
+        {
+            var delayTask = Task.Delay(DelayBetweenConcurrentDemoUsageRetrievalAttempts);
+            var usagesTasks = demos.Select(d => GetLastUsageAsync(d));
+            var usages = await Task.WhenAll(usagesTasks);
+            await delayTask;
+            return usages;
+        }
+
+        private async Task UpdateDemosAsync(IEnumerable<DemoLastUsageRetrieveAttempt> usages)
+        {
+            var usableUsages = usages
+                .Where(a => a.Exception == null)
+                .Where(a => a.LastUsedOn.HasValue)
+                .Select(a => new DemoDeletionSchedule
+                {
+                    Demo = a.Demo,
+                    DeletionScheduledOn = _deletionCalculator.GetDeletionDate(a.Demo, a.LastUsedOn.Value)
+                }).ToList();
+
+            await _demosStore.UpdateDeletionScheduleAsync(usableUsages);
         }
 
         private async Task DeleteAllAsync(IEnumerable<DemoCleanupInfo> info)
@@ -65,40 +123,58 @@ namespace Instances.Application.Demos.Deletion
             {
                 await _ccDataService.DeleteInstancesAsync(batch.Value, batch.Key, $"/api/demos/deletion-report/{batch.Key}");
             }
+
+            await _demosStore.DeleteAsync(info.Select(i => i.Demo));
+            await _instancesStore.DeleteForDemoAsync(info.Select(i => i.Demo.Instance));
         }
 
-        private Task ReportCleanupIntentionsAsync(IEnumerable<DemoCleanupInfo> info)
+        private Task ReportCleanupIntentionsAsync(IEnumerable<DemoCleanupInfo> info, bool isDryRun)
         {
+            var displayName = isDryRun ? "Suppression des démos à blanc" : "Suppression des démos";
             return _emailService.SendAsync
             (
-                new SenderForm { DisplayName = "Suppression des démos" },
+                new SenderForm { DisplayName = displayName },
                 RecipientForm.FromContact(EmailContact.CloudControl),
                 _demoEmails.GetIntentEmail(_timeProvider.Today(), info).Content
             );
         }
 
-        private async Task<DemoCleanupInfo> GetUpdatedCleanupInfoAsync(Demo demo, DateTime demoDeletionSchedule)
+        private async Task<IEnumerable<DemoCleanupInfo>> GetCleanupInfoAsync
+        (
+            IEnumerable<DemoLastUsageRetrieveAttempt> attempts,
+            DateTime today
+        )
+        {
+            return attempts.Select
+            (
+                a => a.Exception != null
+                    ? new DemoCleanupInfo(a.Demo, a.Exception, DemoState.ErrorAtStateEvaluation)
+                    : new DemoCleanupInfo(a.Demo, today)
+            );
+        }
+
+        private async Task<DemoLastUsageRetrieveAttempt> GetLastUsageAsync(Demo demo)
         {
             try
             {
-                await UpdateDeletionScheduleAsync(demo);
-                return new DemoCleanupInfo(demo, demoDeletionSchedule);
+                var latestConnection = await _sessionLogsService.GetLatestAsync(demo.Href);
+                var lastUsedOn = latestConnection == new DateTime(0001, 1, 1)
+                    ? demo.CreatedAt
+                    : latestConnection;
+
+                return new DemoLastUsageRetrieveAttempt { Demo = demo, LastUsedOn = lastUsedOn };
             }
             catch (Exception e)
             {
-                return new DemoCleanupInfo(demo, e, DemoState.ErrorAtStateEvaluation);
+                return new DemoLastUsageRetrieveAttempt { Demo = demo, Exception = e};
             }
         }
 
-        private async Task UpdateDeletionScheduleAsync(Demo demo)
+        private class DemoLastUsageRetrieveAttempt
         {
-            var latestConnection = await _sessionLogsService.GetLatestAsync(demo.Href);
-            var latestDemoUsage = latestConnection == new DateTime(0001, 1, 1)
-                ? demo.CreatedAt
-                : latestConnection;
-
-            var shouldBeDeletedAt = _deletionCalculator.GetDeletionDate(demo, latestDemoUsage);
-            await _demosStore.UpdateDeletionScheduleAsync(demo, shouldBeDeletedAt);
+            public Demo Demo { get; set; }
+            public DateTime? LastUsedOn { get; set; }
+            public Exception Exception { get; set; }
         }
     }
 }
